third_party/WebKit/LayoutTests/fast/forms/number/number-change-event-by-readonly.html

<script src="../../../resources/js-test.js"></script>
<script>
description('Read-only number input should not dispatch a change event.');
var changeEventCounter = 0;
function handleChange() {
    ++changeEventCounter;
}
var parent = document.createElement('div');
parent.innerHTML = '<input type="number" value="0" onchange="handleChange()" readonly>';
shouldBe('changeEventCounter', '0');
</script>
