third_party/blink/web_tests/platform/generic/fast/forms/number/number-change-event-by-readonly-expected.txt

Read-only number input should not dispatch a change event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS changeEventCounter is 0
PASS successfullyParsed is true

TEST COMPLETE
